+++ include/metalang99.h
@@ -10,10 +10,10 @@
 #include <metalang99/choice.h>
 #include <metalang99/control.h>
 #include <metalang99/eval.h>
+#include <metalang99/gen.h>
 #include <metalang99/lang.h>
 #include <metalang99/list.h>
 #include <metalang99/maybe.h>
-#include <metalang99/misc.h>
 #include <metalang99/nat.h>
 #include <metalang99/tuple.h>
 #include <metalang99/util.h>