+++ include/metalang99/util.h
@@ -6,8 +6,6 @@
 #ifndef ML99_UTIL_H
 #define ML99_UTIL_H
 
-#include <metalang99/priv/compiler_specific.h>
-
 #include <metalang99/lang.h>
 
 // Desugaring {
@@ -125,85 +123,6 @@
 #define ML99_consume(...) ML99_call(ML99_consume, __VA_ARGS__)
 
 /**
- * Puts provided arguments into braces.
- *
- * Some code formatters behave strangely when encounter braces in macros (in different places) --
- * this is why this macro exists.
- *
- * # Examples
- *
- * @code
- * #include <metalang99/util.h>
- *
- * // { int a, b, c; }
- * ML99_braced(v(int a, b, c;))
- * @endcode
- */
-#define ML99_braced(...) ML99_call(ML99_braced, __VA_ARGS__)
-
-/**
- * Generates a type definition.
- *
- * # Examples
- *
- * @code
- * #include <metalang99/util.h>
- *
- * // typedef struct { int x, y; } Point;
- * ML99_typedef(v(Point), v(struct { int x, y; }))
- * @endcode
- */
-#define ML99_typedef(ident, ...) ML99_call(ML99_typedef, ident, __VA_ARGS__)
-
-/**
- * Generates a C structure.
- *
- * # Examples
- *
- * @code
- * #include <metalang99/util.h>
- *
- * // struct Point { int x, y; }
- * ML99_struct(v(Point), v(int x, y;))
- * @endcode
- */
-#define ML99_struct(ident, ...) ML99_call(ML99_struct, ident, __VA_ARGS__)
-
-/**
- * Generates an anonymous C structure.
- *
- * # Examples
- *
- * @code
- * #include <metalang99/util.h>
- *
- * // struct { int x, y; }
- * ML99_struct(v(int x, y;))
- * @endcode
- */
-#define ML99_anonStruct(...) ML99_call(ML99_anonStruct, __VA_ARGS__)
-
-/**
- * The same as #ML99_struct but generates a union.
- */
-#define ML99_union(ident, ...) ML99_call(ML99_union, ident, __VA_ARGS__)
-
-/**
- * The same as #ML99_anonStruct but generates a union.
- */
-#define ML99_anonUnion(...) ML99_call(ML99_anonUnion, __VA_ARGS__)
-
-/**
- * The same as #ML99_struct but generates an enumeration.
- */
-#define ML99_enum(ident, ...) ML99_call(ML99_enum, ident, __VA_ARGS__)
-
-/**
- * The same as #ML99_anonStruct but generates an enumeration.
- */
-#define ML99_anonEnum(...) ML99_call(ML99_anonEnum, __VA_ARGS__)
-
-/**
  * A convenience macro to emphasise that your metafunction expands to more than one term.
  *
  * This macro just expands to provided arguments.
@@ -230,8 +149,6 @@
  * // ABC123
  * ML99_CAT_PRIMITIVE(ABC, 123)
  * @endcode
- *
- * @note This macro does not expand to a Metalang99 term.
  */
 #define ML99_CAT_PRIMITIVE(x, y) x##y
 
@@ -249,77 +166,8 @@
  * // "ABC"
  * ML99_STRINGIFY_PRIMITIVE(ABC)
  * @endcode
- *
- * @note This macro does not expand to a Metalang99 term.
  */
 #define ML99_STRINGIFY_PRIMITIVE(...) #__VA_ARGS__
-
-/**
- * Introduces the variable definition @p var_def to a statement right after its invocation.
- *
- * An invocation of #ML99_INTRODUCE_VAR_TO_STMT together with a statement right after it forms a
- * single statement.
- *
- * This macro is useful when you want to generate a sequence of variable definitions inside your
- * macro, but at the same time stick to the following syntax:
- *
- * @code
- * macro(a, b, c) { ... }
- * @endcode
- *
- * Provided that `a`, `b`, and `c` stand for the identifiers of the defined variables, they will be
- * visible only inside a user-supplied statement `{ ... }` and not outside of it.
- *
- * # Example
- *
- * @code
- * #include <metalang99/util.h>
- *
- * for (int i = 0; i < 10; i++)
- *     ML99_INTRODUCE_VAR_TO_STMT(double x = 5.0)
- *     ML99_INTRODUCE_VAR_TO_STMT(double y = 7.0)
- *         printf("i = %d, x = %f, y = %f\n", i, x, y);
- * @endcode
- *
- * @note This macro does not expand to a Metalang99 term.
- */
-#define ML99_INTRODUCE_VAR_TO_STMT(var_def)                                                        \
-    ML99_CLANG_PRAGMA("clang diagnostic push")                                                     \
-    ML99_CLANG_PRAGMA("clang diagnostic ignored \"-Wshadow\"")                                     \
-    for (int metalang99_priv_INTRODUCE_VAR_TO_STMT_break = 0;                                      \
-         metalang99_priv_INTRODUCE_VAR_TO_STMT_break != 1;)                                        \
-        for (var_def; metalang99_priv_INTRODUCE_VAR_TO_STMT_break != 1;                            \
-             metalang99_priv_INTRODUCE_VAR_TO_STMT_break = 1)                                      \
-            ML99_CLANG_PRAGMA("clang diagnostic pop")
-
-/**
- * Suppresses the "unused X" warning right before a statement after its invocation.
- *
- * An invocation of #ML99_SUPPRESS_UNUSED_BEFORE_STMT together with a statement right after it
- * forms a single statement.
- *
- * # Example
- *
- * @code
- * #include <metalang99/util.h>
- *
- * int x, y;
- *
- * for(;;)
- *     ML99_SUPPRESS_UNUSED_BEFORE_STMT(x)
- *     ML99_SUPPRESS_UNUSED_BEFORE_STMT(y)
- *         puts("abc");
- * @endcode
- *
- * @note This macro does not expand to a Metalang99 term.
- */
-#define ML99_SUPPRESS_UNUSED_BEFORE_STMT(expr)                                                     \
-    ML99_CLANG_PRAGMA("clang diagnostic push")                                                     \
-    ML99_CLANG_PRAGMA("clang diagnostic ignored \"-Wshadow\"")                                     \
-    for (int metalang99_priv_SUPPRESS_UNUSED_BEFORE_STMT_break = 0;                                \
-         ((void)(expr), metalang99_priv_SUPPRESS_UNUSED_BEFORE_STMT_break != 1);                   \
-         metalang99_priv_SUPPRESS_UNUSED_BEFORE_STMT_break = 1)                                    \
-        ML99_CLANG_PRAGMA("clang diagnostic pop")
 
 /**
  * Tells whether @p ident belongs to a set of identifiers defined by @p prefix.
@@ -345,8 +193,6 @@
  * // 0
  * ML99_DETECT_IDENT(FOO_, z)
  * @endcode
- *
- * @note This macro does not expand to a Metalang99 term.
  */
 #define ML99_DETECT_IDENT(prefix, ident) ML99_PRIV_IS_TUPLE(ML99_CAT(prefix, ident))
 
@@ -370,23 +216,15 @@
 #ifndef DOXYGEN_IGNORE
 
 // Implementation {
-#define ML99_catEval_IMPL(x, y)       x##y
-#define ML99_cat_IMPL(x, y)           v(ML99_CAT(x, y))
-#define ML99_stringify_IMPL(...)      v(ML99_STRINGIFY(__VA_ARGS__))
-#define ML99_empty_IMPL()             v(ML99_EMPTY())
-#define ML99_id_IMPL(...)             v(ML99_ID(__VA_ARGS__))
-#define ML99_const_IMPL(x, _a)        v(x)
-#define ML99_flip_IMPL(f)             ML99_appl_IMPL(ML99_PRIV_flip, f)
-#define ML99_PRIV_flip_IMPL(f, a, b)  ML99_appl2_IMPL(f, b, a)
-#define ML99_consume_IMPL(...)        v(ML99_CONSUME(__VA_ARGS__))
-#define ML99_braced_IMPL(...)         v({__VA_ARGS__})
-#define ML99_typedef_IMPL(ident, ...) v(typedef __VA_ARGS__ ident;)
-#define ML99_struct_IMPL(ident, ...)  v(struct ident{__VA_ARGS__})
-#define ML99_anonStruct_IMPL(...)     v(struct {__VA_ARGS__})
-#define ML99_union_IMPL(ident, ...)   v(union ident{__VA_ARGS__})
-#define ML99_anonUnion_IMPL(...)      v(union {__VA_ARGS__})
-#define ML99_enum_IMPL(ident, ...)    v(enum ident{__VA_ARGS__})
-#define ML99_anonEnum_IMPL(...)       v(enum {__VA_ARGS__})
+#define ML99_catEval_IMPL(x, y)      x##y
+#define ML99_cat_IMPL(x, y)          v(ML99_CAT(x, y))
+#define ML99_stringify_IMPL(...)     v(ML99_STRINGIFY(__VA_ARGS__))
+#define ML99_empty_IMPL()            v(ML99_EMPTY())
+#define ML99_id_IMPL(...)            v(ML99_ID(__VA_ARGS__))
+#define ML99_const_IMPL(x, _a)       v(x)
+#define ML99_flip_IMPL(f)            ML99_appl_IMPL(ML99_PRIV_flip, f)
+#define ML99_PRIV_flip_IMPL(f, a, b) ML99_appl2_IMPL(f, b, a)
+#define ML99_consume_IMPL(...)       v(ML99_CONSUME(__VA_ARGS__))
 
 #if defined(__GNUC__) && !defined(__clang__)
 #define ML99_PRIV_GCC_PRAGMA(str) _Pragma(str)
@@ -402,22 +240,14 @@
 // }
 
 // Arity specifiers {
-#define ML99_catEval_ARITY    2
-#define ML99_cat_ARITY        2
-#define ML99_stringify_ARITY  1
-#define ML99_empty_ARITY      1
-#define ML99_id_ARITY         1
-#define ML99_const_ARITY      2
-#define ML99_flip_ARITY       1
-#define ML99_consume_ARITY    1
-#define ML99_braced_ARITY     1
-#define ML99_typedef_ARITY    2
-#define ML99_struct_ARITY     2
-#define ML99_anonStruct_ARITY 1
-#define ML99_union_ARITY      2
-#define ML99_anonUnion_ARITY  1
-#define ML99_enum_ARITY       2
-#define ML99_anonEnum_ARITY   1
+#define ML99_catEval_ARITY   2
+#define ML99_cat_ARITY       2
+#define ML99_stringify_ARITY 1
+#define ML99_empty_ARITY     1
+#define ML99_id_ARITY        1
+#define ML99_const_ARITY     2
+#define ML99_flip_ARITY      1
+#define ML99_consume_ARITY   1
 
 #define ML99_PRIV_flip_ARITY 3
 // }